requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
langchain-text-splitters
python-dotenv
pypdf
docx2txt
chromadb
huggingface-hub
sentence-transformers
bitsandbytes
accelerate
requests